fixes tests cleanup
